Adds historical statistics for energy utilization (#16)
Adds a number of new capabilities:
* Support for MFA login
* Auto detection of location IDs, and support for multiple locations on
a single account
* Support for "Sub Account" logins
* Ingestion of statistics for hourly data

WARNING: this will orphan the existing entity - as the location ID is
now part of the entity ID. I couldn't find a way to rename the sensor
entity. There are ways to rename the config entity - but those didn't
seem to apply here.

While this works for me - I would really appreciate some one else trying
it with a different SmartHub Provider.

Closes https://github.com/gagata/ha-smarthub-energy-sensor/issues/14

---------

Co-authored-by: Alex Eiser <alexe@cranksecurity.com>